+++ app/Http/Controllers/VehicleModelController.php
@@ -26,7 +26,7 @@ class VehicleModelController extends Controller
             return view('components.vehicleModels.table', ['vehicleModels' => $vehicleModels])->render();
         }
 
-        return view('backend.vehicleModels.index', compact('vehicleModels'));
+        return view('backend.vehicleModels.index', ['title' => 'Vehicle Models'], compact('vehicleModels'));
     }
 
     public function store(Request $request)
@@ -35,7 +35,7 @@ class VehicleModelController extends Controller
             $request->validate([
                 'name' => 'required|string|max:50|unique:vehicle_models,name',
                 'manufacturer' => 'required|string|max:50',
-                'engine_cc' => 'required|integer',
+                'engine_cc' => 'required|numeric',
                 'fuel_capacity' => 'required|numeric',
                 'payload_capacity' => 'required|numeric',
                 'body_length' => 'nullable|numeric',
@@ -69,7 +69,7 @@ class VehicleModelController extends Controller
             $data = $request->validate([
                 'name' => 'required|string|max:50',
                 'manufacturer' => 'required|string|max:50',
-                'engine_cc' => 'required|integer',
+                'engine_cc' => 'required|numeric',
                 'fuel_capacity' => 'required|numeric',
                 'payload_capacity' => 'required|numeric',
                 'body_length' => 'nullable|numeric',
@@ -92,6 +92,9 @@ class VehicleModelController extends Controller
 
     public function destroy(VehicleModel $vehicleModel)
     {
+        if ($vehicleModel->vehicles()->exists()) {
+            return redirect()->back()->with('error', 'Vehicle Model cannot be deleted as it is associated with vehicles.');
+        }
         $vehicleModel->delete();
 
         return redirect()->back()->with('success', 'Vehicle Model deleted successfully!');
